ai-chatbot/backend/requirements.txt
flask
flask-cors
openai
numpy
PyPDF2
python-dotenv
